Doesn't build yet due to BLayoutBuilder api changes but that's for a next commit.


git-svn-id: file:///srv/svn/repos/haiku/haiku/branches/developer/aldeck/tracker_layout@39030 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/TerminalLine.h
@@ -12,7 +12,7 @@
 
 struct TerminalCell {
 	UTF8Char		character;
-	uint16			attributes;
+	uint32			attributes;
 };
 
 
@@ -30,7 +30,7 @@ struct TerminalLine {
 
 
 struct AttributesRun {
-	uint16	attributes;
+	uint32	attributes;
 	uint16	offset;			// character offset
 	uint16	length;			// length of the run in characters
 };
